+++ src/lib/io_test.rs
@@ -1,7 +1,7 @@
 use super::*;
 
 #[test]
-fn create_text_file_and_delete() {
+fn create_text_file_read_and_delete() {
     let file = create_text_file("test\nend", ".testfile");
     assert!(file.ends_with(".testfile"));
 
@@ -9,7 +9,12 @@ fn create_text_file_and_delete() {
     let mut text = String::new();
     file_obj.read_to_string(&mut text).unwrap();
 
+    let mut file_path = PathBuf::new();
+    file_path.push(&file);
+    let read_text = read_text_file(&file_path);
+
     delete_file(&file);
 
     assert_eq!("test\nend".to_string(), text);
+    assert_eq!(read_text, text);
 }