+++ src/lib/scriptengine/mod_test.rs
@@ -1,5 +1,50 @@
 use super::*;
 use crate::test;
+use crate::types::FileScriptValue;
+
+#[test]
+fn get_script_text_vector() {
+    let output = get_script_text(&ScriptValue::Text(vec![
+        "line 1".to_string(),
+        "line 2".to_string(),
+    ]))
+    .join("\n");
+
+    assert_eq!(output, "line 1\nline 2");
+}
+
+#[test]
+fn get_script_text_file() {
+    let file_info = FileScriptValue {
+        file: "src/lib/test/test_files/text_file.txt".to_string(),
+        absolute_path: None,
+    };
+    let output = get_script_text(&ScriptValue::File(file_info)).join("\n");
+
+    assert_eq!(output, "text 1\ntext 2");
+}
+
+#[test]
+fn get_script_text_file_relative() {
+    let file_info = FileScriptValue {
+        file: "src/lib/test/test_files/text_file.txt".to_string(),
+        absolute_path: Some(false),
+    };
+    let output = get_script_text(&ScriptValue::File(file_info)).join("\n");
+
+    assert_eq!(output, "text 1\ntext 2");
+}
+
+#[test]
+fn get_script_text_file_absolute() {
+    let file_info = FileScriptValue {
+        file: "${CARGO_MAKE_WORKING_DIRECTORY}/src/lib/test/test_files/text_file.txt".to_string(),
+        absolute_path: Some(true),
+    };
+    let output = get_script_text(&ScriptValue::File(file_info)).join("\n");
+
+    assert_eq!(output, "text 1\ntext 2");
+}
 
 #[test]
 fn get_engine_type_no_runner_no_script() {
@@ -13,7 +58,7 @@ fn get_engine_type_no_runner_no_script() {
 #[test]
 fn get_engine_type_no_runner() {
     let mut task = Task::new();
-    task.script = Some(vec!["test".to_string()]);
+    task.script = Some(ScriptValue::Text(vec!["test".to_string()]));
 
     let output = get_engine_type(&task);
 
@@ -34,7 +79,7 @@ fn get_engine_type_no_script() {
 fn get_engine_type_runner_no_extension() {
     let mut task = Task::new();
     task.script_runner = Some("@bad".to_string());
-    task.script = Some(vec!["test".to_string()]);
+    task.script = Some(ScriptValue::Text(vec!["test".to_string()]));
 
     let output = get_engine_type(&task);
 
@@ -45,7 +90,7 @@ fn get_engine_type_runner_no_extension() {
 fn get_engine_type_rust() {
     let mut task = Task::new();
     task.script_runner = Some("@rust".to_string());
-    task.script = Some(vec!["test".to_string()]);
+    task.script = Some(ScriptValue::Text(vec!["test".to_string()]));
 
     let output = get_engine_type(&task);
 
@@ -56,7 +101,7 @@ fn get_engine_type_rust() {
 fn get_engine_type_shell_to_batch() {
     let mut task = Task::new();
     task.script_runner = Some("@shell".to_string());
-    task.script = Some(vec!["test".to_string()]);
+    task.script = Some(ScriptValue::Text(vec!["test".to_string()]));
 
     let output = get_engine_type(&task);
 
@@ -68,7 +113,7 @@ fn get_engine_type_generic() {
     let mut task = Task::new();
     task.script_runner = Some("test1".to_string());
     task.script_extension = Some("test2".to_string());
-    task.script = Some(vec!["test".to_string()]);
+    task.script = Some(ScriptValue::Text(vec!["test".to_string()]));
 
     let output = get_engine_type(&task);
 
@@ -78,7 +123,7 @@ fn get_engine_type_generic() {
 #[test]
 fn invoke_no_runner() {
     let mut task = Task::new();
-    task.script = Some(vec!["echo test".to_string()]);
+    task.script = Some(ScriptValue::Text(vec!["echo test".to_string()]));
 
     let output = invoke(&task, &vec![]);
 
@@ -109,7 +154,7 @@ fn invoke_os_runner() {
     if !test::is_windows_on_travis_ci() {
         let mut task = Task::new();
         task.script_runner = Some(test::get_os_runner());
-        task.script = Some(vec!["echo test".to_string()]);
+        task.script = Some(ScriptValue::Text(vec!["echo test".to_string()]));
 
         let output = invoke(&task, &vec![]);
 
@@ -122,7 +167,9 @@ fn invoke_rust_runner() {
     if test::should_test(false) {
         let mut task = Task::new();
         task.script_runner = Some("@rust".to_string());
-        task.script = Some(vec!["fn main() {println!(\"test\");}".to_string()]);
+        task.script = Some(ScriptValue::Text(vec![
+            "fn main() {println!(\"test\");}".to_string()
+        ]));
 
         let output = invoke(&task, &vec![]);
 
@@ -136,7 +183,9 @@ fn invoke_rust_runner_error() {
     if test::should_test(true) {
         let mut task = Task::new();
         task.script_runner = Some("@rust".to_string());
-        task.script = Some(vec!["fn main() {bad!(\"test\");}".to_string()]);
+        task.script = Some(ScriptValue::Text(vec![
+            "fn main() {bad!(\"test\");}".to_string()
+        ]));
 
         let output = invoke(&task, &vec![]);
 
@@ -148,7 +197,7 @@ fn invoke_rust_runner_error() {
 fn invoke_shell_to_batch_runner() {
     let mut task = Task::new();
     task.script_runner = Some("@shell".to_string());
-    task.script = Some(vec!["echo test".to_string()]);
+    task.script = Some(ScriptValue::Text(vec!["echo test".to_string()]));
 
     let output = invoke(&task, &vec![]);
 
@@ -160,7 +209,7 @@ fn invoke_shell_to_batch_runner() {
 fn invoke_shell_to_batch_runner_error() {
     let mut task = Task::new();
     task.script_runner = Some("@shell".to_string());
-    task.script = Some(vec!["exit 1".to_string()]);
+    task.script = Some(ScriptValue::Text(vec!["exit 1".to_string()]));
 
     let output = invoke(&task, &vec![]);
 
@@ -172,7 +221,7 @@ fn invoke_generic_runner() {
     let mut task = Task::new();
     task.script_runner = Some(test::get_os_runner());
     task.script_extension = Some(test::get_os_extension());
-    task.script = Some(vec!["echo test".to_string()]);
+    task.script = Some(ScriptValue::Text(vec!["echo test".to_string()]));
 
     let output = invoke(&task, &vec![]);
 
@@ -185,7 +234,7 @@ fn invoke_generic_runner_error() {
     let mut task = Task::new();
     task.script_runner = Some(test::get_os_runner());
     task.script_extension = Some(test::get_os_extension());
-    task.script = Some(vec!["exit 1".to_string()]);
+    task.script = Some(ScriptValue::Text(vec!["exit 1".to_string()]));
 
     let output = invoke(&task, &vec![]);